% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_status.R
\name{fire_strange}
\alias{fire_strange}
\title{Title}
\usage{
fire_strange(status_file, verbose = options()[["verbose"]])
}
\arguments{
\item{status_file}{(chr) status file}
}
\value{
Value
}
\description{
Title
}
