% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_video_funs.R
\name{get_frame_path}
\alias{get_frame_path}
\title{Title}
\usage{
get_frame_path(folder, frame_id, pid)
}
\arguments{
\item{folder}{(chr) folder}

\item{frame_id}{(chr) camera frame (0, 1, ...)}

\item{pid}{(chr) Personal ID}
}
\value{
Value
}
\description{
Title
}
