% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_video_funs.R
\name{get_video_path}
\alias{get_video_path}
\title{Title}
\usage{
get_video_path(folder, pid)
}
\arguments{
\item{folder}{(chr) folder}

\item{pid}{(chr) Personal ID}
}
\value{
Value
}
\description{
Title
}
