% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobed.shiny-package.R
\docType{package}
\name{iobed.shiny-package}
\alias{iobed.shiny}
\alias{iobed.shiny-package}
\title{iobed.shiny: I.O.BED UBEP data collection}
\description{
Data collection interface for devices used by UBEP for the I.O.BED project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/iobed.shiny}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/iobed.shiny/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it}

}
\keyword{internal}
