% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_video_funs.R
\name{is_status}
\alias{is_status}
\title{Title}
\usage{
is_status(status_file, status)
}
\arguments{
\item{status_file}{(chr) status file}

\item{status}{(chr) status to detect}
}
\value{
Value
}
\description{
Title
}
