% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_status.R
\name{set_status}
\alias{set_status}
\title{Title}
\usage{
set_status(status_file, status_msg)
}
\arguments{
\item{status_file}{(chr) status file}

\item{status_msg}{(chr) message}
}
\value{
Value
}
\description{
Title
}
