% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_status.R
\name{fire_running}
\alias{fire_running}
\title{Title}
\usage{
fire_running(status_file, perc = NULL)
}
\arguments{
\item{status_file}{(chr) status file}

\item{perc}{(num) complete rate}
}
\value{
Value
}
\description{
Title
}
